Label pairs added, new project structure, test refactoring
