+++ metrics-scm-1.rockspec
@@ -46,8 +46,8 @@ build = {
         ['metrics.default_metrics.tarantool.slab']       = 'metrics/default_metrics/tarantool/slab.lua',
         ['metrics.default_metrics.tarantool.spaces']     = 'metrics/default_metrics/tarantool/spaces.lua',
         ['metrics.default_metrics.tarantool.system']     = 'metrics/default_metrics/tarantool/system.lua',
-        ['metrics.cartridge_metrics']                    = 'metrics/cartridge_metrics/init.lua',
-        ['metrics.cartridge_metrics.issues']             = 'metrics/cartridge_metrics/issues.lua',
+        ['metrics.default_metrics.cartridge']            = 'metrics/default_metrics/cartridge/init.lua',
+        ['metrics.default_metrics.cartridge.issues']     = 'metrics/default_metrics/cartridge/issues.lua',
         ['metrics.psutils.cpu']                          = 'metrics/psutils/cpu.lua',
         ['metrics.psutils.psutils_linux']                = 'metrics/psutils/psutils_linux.lua',
         ['metrics.tarantool.luajit']                     = 'metrics/tarantool/luajit.lua',
